DEVELOP - Refactor game and round modules by changing roundCount type to int16 for consistency, and introduce new market-related types and repository for market management. Add initial market handler and model definitions to support market creation functionality.

+++ modules/rounds/repository.go
@@ -42,7 +42,7 @@ func Update(round *Round) error {
 	return nil
 }
 
-func GetTodaysRoundCount() int64 {
+func GetTodaysRoundCount() int16 {
 	coll := getRoundsCollection()
 
 	filter := bson.M{
@@ -60,12 +60,12 @@ func GetTodaysRoundCount() int64 {
 
 	// Extract the last 5 characters from display_id and convert to int64
 	displayIDSuffix := latestRound.DisplayID[len(latestRound.DisplayID)-5:]
-	count := int64(0)
+	count := int16(0)
 
 	// Parse the suffix to get the count
 	for _, char := range displayIDSuffix {
 		if char >= '0' && char <= '9' {
-			count = count*10 + int64(char-'0')
+			count = count*10 + int16(char-'0')
 		}
 	}
